Modificacion del menu platillos y buscador en el home
