+++ src/components/sprint2/MenuPlatillos/menuPlatillos.css
@@ -1,18 +1,8 @@
-.estilo-search{
 
-}
-
-.estilo-autocompletable{
-    position: absolute;
-    width: 40%;
-    margin-top: -80px;
-    margin-left: 54%;
-}
-
-.estilo-button{
+.estilo-button2{
     position: absolute  ;
     width: 20%;
     height: 37px;
-    margin-top: -80px;
+    margin-top: -295px;
     margin-left: 91.6%;
 }
